requirements.txt
elastic_transport==8.4.0
elasticsearch==8.7.0
python_dateutil==2.8.2
PyYAML==6.0
PyYAML==6.0
